JAVA-2564: Refactor Mongo class to create a separate class for the ServerSession pool
